desktop and window (1st desktop and mouse)

+++ src/drivers/mouse.cpp
@@ -76,7 +76,7 @@ uint32_t MouseDriver::HandleInterrupt(uint32_t esp)
     {
         if(buffer[1] != 0 || buffer[2] != 0)
         {
-            handler->OnMouseMove(buffer[1], -buffer[2]);
+            handler->OnMouseMove((int)buffer[1], -((int)buffer[2]));
         }
         
         for (uint8_t i = 0; i < 3; i++)
